+++ src/pages/add.vue
@@ -37,7 +37,6 @@
 
 <script>
 import { mapState, mapMutations } from 'vuex'
-import mpButton from 'mpvue-weui/src/button'
 
 export default {
   data () {
@@ -48,9 +47,6 @@ export default {
       thing: '',
       place: ''
     }
-  },
-  components: {
-    mpButton
   },
   computed: {
     ...mapState([
